rootfs-postcommands: Cope with empty IMAGE_LINK_NAME in write_image_test_data
Ensure that we don't create an image test data symlink named
".testdata.json" when IMAGE_LINK_NAME is empty.

+++ meta/classes/rootfs-postcommands.bbclass
@@ -328,7 +328,7 @@ python write_image_test_data() {
     searchString = "%s/"%(d.getVar("TOPDIR")).replace("//","/")
     export2json(d, testdata_name, searchString=searchString, replaceString="")
 
-    if os.path.exists(testdata_name):
+    if os.path.exists(testdata_name) and link_name:
         testdata_link = os.path.join(deploy_dir, "%s.testdata.json" % link_name)
         if os.path.lexists(testdata_link):
             os.remove(testdata_link)
